changes in models

+++ src/main/java/com/ulacit/matriculas/matriculasulacit/Repository/PersonaRepository.java
@@ -4,7 +4,7 @@ import com.ulacit.matriculas.matriculasulacit.Modelos.Persona;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonaRepository extends JpaRepository<Persona, Long> {
+public interface PersonaRepository extends JpaRepository<Persona, Integer> {
     List<Persona> findByDeleted(Boolean deleted);
     Persona findByIdPersonaInAndDeletedIn(int idPersona, Boolean deleted);
 }
